requirements.txt
PyYAML>=3.12
injector>=0.14.1
coloredlogs>=10.0
Jinja2>=2.6